requirements.txt
beautifulsoup4==4.12.3
boto3==1.35.66
botocore==1.35.66
bs4==0.0.2
certifi==2024.8.30
charset-normalizer==3.4.0
idna==3.10
jmespath==1.0.1
numpy==2.1.3
pandas==2.2.3
pip==24.3.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
s3transfer==0.10.4
six==1.16.0
soupsieve==2.6
tzdata==2024.2
urllib3==2.2.3
vaderSentiment==3.3.2
matplotlib==3.9.3
keras==3.7.0
tensorflow==2.18.0
SQLAlchemy==2.0.36
PyMySql==1.1.1
cryptography==44.0.0
